lambda-ocr-face/local-test/requirements.txt
# Core
python-dotenv>=1.0.0
boto3>=1.26.0

# Database
psycopg[binary]>=3.1.0
pgvector>=0.2.0

# OCR
paddlepaddle==3.0.0
paddleocr==2.9.1

# Face Detection
insightface>=0.7.0
onnxruntime>=1.15.0

# Image Processing
opencv-python-headless>=4.5.0
Pillow>=10.0.0
numpy>=1.24.0